add modified files
