+++ newproj/api/__init__.py
@@ -4,7 +4,6 @@ app = Flask (__name__)
 
 from api import indexpage
 from api import dummy_data
-from api import users
 from api import requests
 from api import models
 from api import user_signin
